+++ app/src/main/java/com/simongellis/vvb/emulator/Audio.kt
@@ -25,7 +25,7 @@ class Audio(emulator: Emulator, settings: Settings) {
         nativePause()
     }
 
-    class Settings(val volume: Int, val bufferSize: Int)
+    class Settings(val volume: Float, val bufferSize: Int)
 
     private external fun nativeConstructor(emulator: Emulator, settings: Settings)
     private external fun nativeDestructor()
